+++ test/runn_rsc_test.dart
@@ -31,6 +31,8 @@ void main() {
     expect(treadmill.canMeasureHeartRate, false);
     expect(treadmill.defaultSport, ActivityType.run);
     expect(treadmill.fourCC, npeRunnFourCC);
+    expect(treadmill.isMultiSport, false);
+    expect(treadmill.shouldSignalStartStop, false);
   });
 
   test('Runn RSC Device interprets flags properly', () async {