+++ test/schwinn_x70_test.dart
@@ -118,7 +118,8 @@ void main() {
         if (record.speed == null) {
           expect(testPair.record.speed, null);
         } else {
-          expect(record.speed!, closeTo(testPair.record.speed!, eps));
+          expect(record.speed!,
+              closeTo(testPair.record.speed!, testPair.record.speed! * workaroundEpsRatio));
         }
         expect(record.cadence, testPair.record.cadence);
         expect(record.heartRate, testPair.record.heartRate);
@@ -248,7 +249,8 @@ void main() {
         if (record.speed == null) {
           expect(testPair.record.speed, null);
         } else {
-          expect(record.speed!, closeTo(testPair.record.speed!, eps));
+          expect(record.speed!,
+              closeTo(testPair.record.speed!, testPair.record.speed! * workaroundEpsRatio));
         }
         expect(record.cadence, testPair.record.cadence);
         expect(record.heartRate, testPair.record.heartRate);